test: fold attestor tests into generator_test
Move the attestor tests into generator_test.go and drop attestor_test.go
so tests track the single generator.go source; prompt_test.go stays split.
