--- pkg/generator/attestor_test.go
@@ -1,65 +1,0 @@
-// Copyright 2025 venslabs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package generator
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	"github.com/venslabs/vens/internal/testutil"
-	"github.com/venslabs/vens/pkg/attestation"
-	"github.com/venslabs/vens/pkg/llm"
-	"github.com/venslabs/vens/pkg/riskconfig"
-)
-
-func testVulns(n int) []Vulnerability {
-	v := make([]Vulnerability, n)
-	for i := range v {
-		v[i] = Vulnerability{VulnID: fmt.Sprintf("CVE-2024-%04d", i), PkgName: "pkg", Title: "t"}
-	}
-	return v
-}
-
-// One evidence batch is recorded per LLM call, so BatchCount tracks the batching.
-func TestGenerator_Attestor_OneBatchPerLLMCall(t *testing.T) {
-	at := attestation.NewBuilder(attestation.Opts{Provider: "mock", Model: "mock"})
-	g, err := New(Opts{LLM: testutil.NewMockLLM(), Config: &riskconfig.Config{}, BatchSize: 10})
-	require.NoError(t, err)
-	g.SetAttestor(at)
-
-	require.NoError(t, g.GenerateRiskScore(context.Background(), testVulns(15), nil))
-	require.Equal(t, 2, at.BatchCount()) // 15 vulns / batch size 10 -> 2 batches
-}
-
-type failingLLM struct{}
-
-func (failingLLM) Generate(context.Context, llm.Request) (string, error) {
-	return "", errors.New("llm down")
-}
-
-// A failed LLM call must not record evidence, or the attestation would carry
-// empty/garbage batches.
-func TestGenerator_Attestor_NoBatchOnLLMError(t *testing.T) {
-	at := attestation.NewBuilder(attestation.Opts{Provider: "mock", Model: "mock"})
-	g, err := New(Opts{LLM: failingLLM{}, Config: &riskconfig.Config{}})
-	require.NoError(t, err)
-	g.SetAttestor(at)
-
-	require.Error(t, g.GenerateRiskScore(context.Background(), testVulns(3), nil))
-	require.Equal(t, 0, at.BatchCount())
-}
